More Error handling and code cleanup
